Upload Previous Project
The files and folders are organized as follows:

The solution file of the whole program is "ImbalancedKeypoints.sln" with its related folders created by Microsoft Visual Studio 2015 named "Debug" and "x64".

Main program and subprograms are stored in its name-based folders: "CellularAutomaton", "ImbalancedKeypoints", "ImbDetector", "MeshMatcher", "MeshSubdivisor", and "MeshSurfaceFitter".

Folders named "CloudFile" and "MeshFile" are used for storing data files in the form of point clouds and meshes respectively.

The experimental comparisons and analysis are held in the folder "Comparison", while "Paper and Material" holds the related materials and academic papers.

In addition, A file named "Environment" consists of the information about how to build up the experimental environment to run the whole program. Moreover, "ImbalancedKeypoints_Release_V_10_Carl.7z" is a package of the environment-independent executable software, which is able to run on each x64 Windows operating system.